// api/src/modules/transfers/api/models/batchTransferResponseV1.ts
/**
 * Transfers API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { TransferResponseV1 } from './transferResponseV1';

export interface BatchTransferResponseV1 {
  id: string;
  state: BatchTransferResponseV1State;
  transferType: BatchTransferResponseV1TransferType;
  batchTransferId: string;
  totalAmount: number;
  metadata: any;
  webhookClientURL?: string;
  transfers: Array<TransferResponseV1>;
}
export enum BatchTransferResponseV1State {
  Pending = 'pending',
  Processing = 'processing',
  Completed = 'completed',
  Failed = 'failed',
}
export enum BatchTransferResponseV1TransferType {
  Regular = 'regular',
  Expedite = 'expedite',
  Ach = 'ach',
  Rtp = 'rtp',
}

export interface BatchTransferResponseV1List {
  items: BatchTransferResponseV1[];
  nextCursor?: string;
}